Size factoring change

+++ src/variables/bencode_object.rs
@@ -176,7 +176,7 @@ impl BencodeVariable for BencodeObject {
 
         let mut res = OrderedMap::<BencodeBytes, Box<dyn BencodeVariable>>::new();//::with_hasher(Default::default());
 
-        while buf[off] != Self::TYPE.suffix() as u8 {
+        while buf[off] != Self::TYPE.suffix() {
             let key = BencodeBytes::decode_with_offset(buf, off);
             off += key.byte_size();
             let type_ = BencodeType::type_by_prefix(buf[off]);
